bug: Fix long text messages not shrinking
When "overflow-wrap: break-word" was added to the chat messages,
it allowed full words to be wrapped onto the next line. However,
this caused long text messages to not shrink within the ChatMessages
component. This caused the ChatSidebar to be hidden.

Below is a representation of the JSX structure (<ChatContainer>
is used for demonstrating the problem):

<Chat>
  <ChatContainer>
    <ChatMessages />
    <ChatInput />
  </ChatContainer>
  <ChatSidebar />
</ChatSidebar>

<Chat> is a flex container. By default flex children have a "min-width: auto;"
This auto is what prevented the <ChatContainer> from shrinking. It didn't
shrink because auto allowed the long text to take up all available screen
width.

The fix was to set the min-width of <ChatContainer> to zero.